Move run_monitor.js to scripts / cleanup rebase
