+++ package.json
@@ -17,7 +17,7 @@
     "dev:ui": "cd ./ui/ && npm install && npm start",
     "dev:server": "npm run build:ui && ./scripts/run_server.sh",
     "dev:server:local": "npm run build:ui && TESTING=true ./scripts/run_server.sh",
-    "build:ui": "cd ./ui/ && npm install && npm run build"
+    "build:ui": "cd ./ui/ && npm install && npm run build",
     "monitor": "node ./scripts/run_monitor.js",
     "server": "node ./src/server.js ./repository"
   },
